refactor out mixin
