--- .eslintrc
@@ -1,38 +1,0 @@
-{
-  "parser": "babel-eslint",
-  "env": {
-    "browser": true,
-    "node": true
-  },
-  "ecmaFeatures": {
-    "jsx": true
-  },
-  "plugins": [
-    "react"
-  ],
-  "rules": {
-    "quotes": [2, "single"],
-    "eol-last": [2],
-    "no-underscore-dangle": [0],
-    "no-cond-assign": 0,
-    "indent": [2, 2],
-    "semi": [2, "always"],
-    "camelcase": 2,
-    "comma-style": [2, "first"],
-    "new-cap": 0,
-    "strict": 2,
-    "no-unused-expressions": 0,
-    "semi-spacing": 0,
-    "consistent-return": 2,
-    "default-case": 2,
-    "radix": 2,
-    "brace-style": 0,
-    "dot-location": [2, "property"],
-    "react/display-name": 1,
-    "react/jsx-no-undef": 2,
-    "react/jsx-quotes": [2, "double"],
-    "react/jsx-uses-react": 2,
-    "react/jsx-uses-vars": 2,
-    "react/react-in-jsx-scope": 2
-  }
-}
